bump env to torch 2.11 / cuda 13 / pyg 2.7 and adapt code
Refresh env.yml: PyTorch 2.4.1 -> 2.11.0 (gpu_cuda130* build), drop
abandoned pytorch-cuda metapackage, install PyG 2.7.0 via pip, lightning
>= 2.4, numpy >= 2, pymatgen < 2026.5 (avoids in-progress pymatgen-core
split), lmdb < 2 (newer lmdb refuses same-process duplicate opens).
Also add wandb / tensorboard / seaborn / pyarrow / tqdm so the full
test suite and CLI scripts work in a fresh env, plus parity with
qtaim_generator so both repos can share one conda env.

Code adaptations:
* `_pymatgen_compat.py` patches `Composition.__setstate__` to map the
  legacy `_natoms` field to `_n_atoms`, letting pickled Compositions
  produced under pymatgen ~2023 load on newer pymatgen. Imported from
  the package __init__.
* Lightning 2.x: replace `precision=16` literals with `"16-mixed"` in 4
  test files; rewrite the precision-string branches in the 6 train /
  eval scripts so a JSON config `"16"` becomes `"16-mixed"` rather than
  int 16.
* PyTorch 2.x: drop the removed `verbose=True` kwarg from all 4
  `ReduceLROnPlateau` call sites.
* pandas 2.x: switch `col.iloc[ind][0]` to `col.iloc[ind, 0]` in
  test_featurizers; the positional-by-label fallback is gone.
* LMDB: share env handles across `LMDBBaseDataset` instances via a
  process-local `_env_cache` (newer lmdb forbids opening the same path
  twice in one process); reset the cache after fork and between pytest
  tests via an autouse fixture in conftest.

Tests: 201 passed, 1 skipped on the new env.

+++ tests/test_lmdb.py
@@ -185,7 +185,7 @@ def test_model_lmdb(device_config):
         enable_progress_bar=True,
         enable_checkpointing=True,
         default_root_dir="./test_save_load/",
-        precision=16,
+        precision="16-mixed",
         log_every_n_steps=1,
     )
 
@@ -224,7 +224,7 @@ def test_model_lmdb_multi_file(device_config):
         enable_progress_bar=True,
         enable_checkpointing=True,
         default_root_dir="./test_save_load/",
-        precision=16,
+        precision="16-mixed",
         log_every_n_steps=1,
     )
